Add Email Registration component and authentication state management
